docs(ecosystem): redesign the page as tables and publish it
The card grid is gone; the page is now three device tables -- Supported
Devices, Data Factory, and Cloud Infrastructure -- all fed by devices.yaml.

The card data, directive, styles, preview page, and intake script are
deleted as unused, and partner_grid.py becomes ecosystem_grid.py. The Git
LFS carve-out added for the card logos is reverted. conf.py and index.rst
publish the page.

docs/AGENTS.md is rewritten to match.
