+++ docs/source/_static/logos/haptikos-dark.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9ab10d45750281337ab32850fe75cb34531af69f74592e91bc9e7fbc4e9c0b65
+size 19641
