+++ docs/source/_static/logos/haptikos.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:39a36d04b236241261c8695b3ce57530ffeb492f923ed68f94805ad8c51762eb
+size 19696
