+++ docs/source/index.rst
@@ -42,6 +42,7 @@ Table of Contents
 
    Isaac Teleop <self>
    overview/architecture
+   overview/ecosystem
 
 .. toctree::
    :maxdepth: 2
